[ci] Build.py: Rename ci_common -> ti_build, add root level entrypoint
